More spam fixing
Can't have reply-to as the sender's email address, so now the emails
come from noreply@. Added a mailto: link within the email to reply
directly to the sender, and a link to the group page.
Also improved html formatting, but I may need to add a plain text
version in due course.